pyproject.toml
[project]
name = "realtime-system"
version = "0.1.0"
description = "Local realtime voice assistant on Apple Silicon using MLX."
requires-python = ">=3.11"
dependencies = [
  "mlx>=0.24.0",
  "mlx-lm>=0.22.0",
  "mlx-audio>=0.3.0",
  "misaki>=0.9.4",
  "phonemizer-fork>=3.3.2",
  "espeakng-loader>=0.2.4",
  "num2words>=0.5.14",
  "spacy>=3.8.4",
  "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
  "numpy>=1.26.0",
  "sounddevice>=0.5.0",
  "webrtcvad-wheels>=2.0.14",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]
